Fix pre-existing beartype annotation errors surfaced at runtime
- Remove beartype-incompatible local variable annotations for bgr_3hw
  and intri_torch in run_dpvo_pipeline (jaxtyping dimension name
  conflicts with beartype's checker cache)
- Fix DPVO.__call__ image param: Float -> UInt8 (input is uint8,
  normalized to float inside the method)

+++ packages/mini-dpvo/mini_dpvo/dpvo.py
@@ -669,7 +669,7 @@ class DPVO:
             indexing="ij",
         )
 
-    def __call__(self, tstamp: int, image: Float[Tensor, "3 ht wd"], intrinsics: Float[Tensor, "4"]) -> None:
+    def __call__(self, tstamp: int, image: UInt8[Tensor, "3 ht wd"], intrinsics: Float[Tensor, "4"]) -> None:
         """Process a new frame: extract patches, predict pose, update and optimize.
 
         This is the main entry point called once per frame.  The full
